+++ fenix/automation/taskcluster/androidTest/flank-arm-legacy-api-tests.yml
@@ -23,15 +23,13 @@ gcloud:
     - notPackage org.mozilla.fenix.screenshots
     - notPackage org.mozilla.fenix.syncintegration
     - notPackage org.mozilla.fenix.experimentintegration
-    - class org.mozilla.fenix.ui.NavigationToolbarTest#visitURLTest
-    - class org.mozilla.fenix.ui.NavigationToolbarTest#goBackTest
-    - class org.mozilla.fenix.ui.NavigationToolbarTest#goForwardTest
+    - class org.mozilla.fenix.ui.MainMenuTest#goBackTest
+    - class org.mozilla.fenix.ui.MainMenuTest#goForwardTest
     - class org.mozilla.fenix.ui.HistoryTest#visitedUrlHistoryTest
     - class org.mozilla.fenix.ui.MainMenuTest#openMainMenuSettingsItemTest
     - class org.mozilla.fenix.ui.SettingsSearchTest#toggleSearchSuggestionsTest
     - class org.mozilla.fenix.ui.CollectionTest#deleteCollectionTest
     - class org.mozilla.fenix.ui.SmokeTest#noHistoryInPrivateBrowsingTest
-    - class org.mozilla.fenix.ui.SmokeTest#verifyBasicNavigationToolbarFunctionality
     - class org.mozilla.fenix.ui.TabbedBrowsingTest#openNewTabTest
     - class org.mozilla.fenix.ui.TabbedBrowsingTest#openNewPrivateTabTest
     - class org.mozilla.fenix.ui.TopSitesTest#verifyOpenTopSiteNormalTab
